Create different pages of the walkthrough separately
